chore: normalize document serverservice usage (#453)
