+++ packages/services/document-service/.dockerignore
