+++ packages/services/document-service/.gitignore
