+++ packages/services/document-service/gradle.properties
