+++ packages/services/document-service/src/main/java/software/uncharted/terarium/documentserver/models/xdd/Extraction.java
