+++ packages/services/document-service/src/main/java/software/uncharted/terarium/documentserver/resources/xdd/DocumentResource.java
@@ -10,7 +10,7 @@ import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-@Authenticated
+
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
 @Tag(name = "XDD Documents REST Endpoint")
