+++ packages/services/document-service/src/main/java/software/uncharted/terarium/documentserver/resources/xdd/ExtractionResource.java
@@ -10,7 +10,6 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
 
-@Authenticated
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
 @Tag(name = "XDD Extraction REST Endpoint")
