+++ packages/services/document-service/src/main/java/software/uncharted/terarium/documentserver/resources/xdd/SetResource.java
@@ -5,25 +5,26 @@ import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
 import software.uncharted.terarium.documentserver.proxies.xdd.DocumentProxy;
 
-
-import javax.ws.rs.*;
+import javax.ws.rs.Consumes;
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-
-@Authenticated
-@Tag(name = "XDD Dictionaries REST Endpoint")
-@Path("/api")
-public class DictionariesResource {
+@Path("/sets")
+@Produces(MediaType.APPLICATION_JSON)
+@Consumes(MediaType.APPLICATION_JSON)
+@Tag(name = "XDD Sets REST Endpoint")
+public class SetResource {
 
 	@RestClient
 	DocumentProxy proxy;
 
 	@GET
 	@Produces(MediaType.APPLICATION_JSON)
-	@Tag(name = "Get available XDD dictionaries")
-	@Path("/dictionaries")
-	public Response getAvailableDictionaries(@QueryParam("all") @DefaultValue("") String all) {
-		return proxy.getAvailableDictionaries(all);
+	@Tag(name = "Get available XDD sets or collections")
+	public Response getAvailableSets() {
+		return proxy.getAvailableSets();
 	}
 }
