+++ packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/proxies/xdd/DocumentProxy.java
@@ -6,7 +6,7 @@ import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-@RegisterRestClient(configKey = "xdd-document-service")
+@RegisterRestClient(configKey = "document-service")
 @Produces(MediaType.APPLICATION_JSON)
 public interface DocumentProxy {
 	@GET
