+++ packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/proxies/xdd/ExtractionProxy.java
@@ -9,7 +9,7 @@ import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-@RegisterRestClient(configKey = "xdd-extraction-service")
+@RegisterRestClient(configKey = "extraction-service")
 @Produces(MediaType.APPLICATION_JSON)
 public interface ExtractionProxy {
 	@GET
