+++ packages/services/hmi-server/src/main/resources/application.properties
@@ -25,10 +25,10 @@ mp.openapi.extensions.smallrye.info.description=REST endpoints for the TERArium 
 %prod.data-service/mp-rest/url=http://${data-service}:3020
 %dev.model-service/mp-rest/url=http://localhost:3010
 %prod.model-service/mp-rest/url=http://${model-service}:3010
-%dev.xdd-document-service/mp-rest/url=http://localhost:3030
-%prod.xdd-document-service/mp-rest/url=http://${document-service}:3030
-%dev.xdd-extraction-service/mp-rest/url=http://localhost:3030
-%prod.xdd-extraction-service/mp-rest/url=http://${document-service}:3030
+%dev.document-service/mp-rest/url=http://localhost:3030
+%prod.document-service/mp-rest/url=http://${document-service}:3030
+%dev.extraction-service/mp-rest/url=http://localhost:3030
+%prod.extraction-service/mp-rest/url=http://${document-service}:3030
 ########################################################################################################################
 # Native container configurations
 ########################################################################################################################
